Be smarter about the direction we split
Use the dimensions of the Vim window to determine the desired split
direction. There are configurable options to make your split direction
more constant.

Add an option to always split.

Add modeline consistent with current formatting.

Update the help. (Also add a better intro and a link to the license.)