Scratch takes filetype for scratch buffer
Instead of using the argument as just a name, take a filetype. You can
still give an invalid filetype as an argument to get the old behavior
except now all scratch buffers are numbered.

The filetype allows us to preset the scratch buffer with a filetype for
syntax, indent, blah blah, goodness.

Instead of an actual filetype, you can pass a dot (.) to use the current
buffer's filetype (the active one when you issue :Scratch).