Add Saveas command to save a scratch buffer
Scratch buffers give errors if you try to save them, so provide a
command to turn off the scratch settings and pass the bang and file
through to built-in saveas.

Saveas command only exists in scratch buffers.

Addresses #2. There's no cmdalias because it's too awkward. (I actually
always use 'save' instead of 'saveas', but 'sav' is also valid. I don't
want to have an explosion of aliases, so I'll leave it as a straight
command for now.)